added obj detect button
added obj detect and detect output button to the list of buttons.  also added img gameobject that will be replaced by a pannel.  The detect handler script is used to control what shows up when the output changes.
However, it currently need to be linked from a normal button to the fancy button before it can work